refactor embedding preprocess code

+++ asreview2-optuna/feature_matrix_scripts/tfidf.py
@@ -2,20 +2,34 @@ from pathlib import Path
 import pickle
 
 import synergy_dataset as sd
-import asreview as asr
+import pandas as pd
+from tqdm import tqdm
 
 from asreview.models.feature_extraction import Tfidf
 
-print("synergy-dataset version", sd.__version__)
-print("asreview version:", asr.__version__)
+FORCE = False
 
-folder_pickle_files = Path("synergy-dataset", "pickles")
+folder_pickle_files = Path("synergy-dataset", "pickles_tfidf")
 folder_pickle_files.mkdir(parents=True, exist_ok=True)
 
-for dataset in sd.iter_datasets():
-    df = dataset.to_frame().reset_index()
+for dataset in tqdm(sd.iter_datasets(), total=26):
+    if dataset.name == "Moran_2021":
+        df = pd.read_csv("../datasets/Moran_2021_corrected_shuffled_raw.csv")
+    else:
+        # Convert dataset to a DataFrame and reset index
+        df = dataset.to_frame().reset_index()
+
+    dataset_name = (
+        dataset.name if dataset.name != "Moran_2021" else "Moran_2021_corrected"
+    )
+    pickle_file_path = folder_pickle_files / f"{dataset_name}.pkl"
+
+    # Check if the pickle file already exists
+    if not FORCE and pickle_file_path.exists():
+        print(f"Skipping {dataset_name}, pickle file already exists.")
+        continue
 
     X = Tfidf().fit_transform(df)
 
-    with open(folder_pickle_files / f"{dataset.name}.pkl", "wb") as f:
+    with open(folder_pickle_files / f"{dataset_name}.pkl", "wb") as f:
         pickle.dump((X, df["label_included"]), f)
